requirements.txt
streamlit>=1.40.0
scikit-learn>=1.5.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-docx>=1.1.0
pydantic>=2.0.0
pytest>=8.0.0
